Fix: UI Scaling Issue Feat: Added Pixel Font
This should fix the UI Scaling issue. I added a pixel font and added raster rendering to it to keep with the style. This is the beginning of the UI update I plan on committing later on tonight before the deadline. Seeing Thor isn't looking at the games just yet and we have a itch page for it. I believe we can continue adding stuff since the MVP is finished.